- perbaiki weight sum (choose courier)

+++ app/src/main/java/com/nyelam/android/data/DoShopProduct.java
@@ -30,7 +30,7 @@ public class DoShopProduct implements Parseable {
     private String productName;
     private String featuredImage;
     private List<String> images;
-    private int weight;
+    private double weight;
     private double specialPrice;
     private double normalPrice;
     private String status;
@@ -138,11 +138,11 @@ public class DoShopProduct implements Parseable {
 //    }
 
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
 
@@ -253,8 +253,8 @@ public class DoShopProduct implements Parseable {
         try {
             if (!obj.isNull(KEY_WEIGHT) && obj.get(KEY_WEIGHT) instanceof String) {
                 setWeight(Integer.valueOf(obj.getString(KEY_WEIGHT)));
-            } else if (!obj.isNull(KEY_WEIGHT) && obj.get(KEY_WEIGHT) instanceof Integer) {
-                setWeight(obj.getInt(KEY_WEIGHT));
+            } else if (!obj.isNull(KEY_WEIGHT) && obj.get(KEY_WEIGHT) instanceof Double) {
+                setWeight(obj.getDouble(KEY_WEIGHT));
             }
         } catch (JSONException e) {e.printStackTrace();}
